fix(infra): isolate Node-only instrumentation from Edge bundle trace (#380)
